requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
beartype==0.18.4
beautifulsoup4==4.12.3
bs4==0.0.2
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
cssselect==1.2.0
fake-headers==1.0.2
fake-useragent==1.5.1
fast_bitrix24==1.7.1
flake8==7.0.0
frozenlist==1.4.1
html5lib==1.1
icontract==2.6.6
idna==3.7
importlib_metadata==7.1.0
lxml==5.2.1
lxml_html_clean==0.1.1
mccabe==0.7.0
more-itertools==10.2.0
multidict==6.0.5
parse==1.20.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyee==11.1.0
pyflakes==3.2.0
pyppeteer==2.0.0
pyquery==2.0.0
requests==2.31.0
requests-cache==1.2.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
url-normalize==1.4.3
urllib3==1.26.18
utils==1.0.2
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
yarl==1.9.4
zipp==3.18.1
